backend/requirements/base.txt
Django>=4.2,<5.0
django-ninja>=1.1.0
pydantic>=1.10.7
django-environ>=0.10.0
dj-database-url>=2.0.0
django-cors-headers>=3.14.0
psycopg[binary]>=3.1.8
celery>=5.3.0
redis>=4.5.0
django-redis>=5.4.0
httpx>=0.24.0
PyJWT>=2.7.0
google-auth-oauthlib>=1.0.0
channels>=3.0.5
channels-redis>=4.0.0
qdrant-client>=1.6.9
mem0ai>=0.1.0
python-dotenv>=1.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0
